fix delete bug, add lihat stok

+++ app/Http/Controllers/KendaraanController.php
@@ -6,6 +6,7 @@ use App\Models\Mobil;
 use App\Models\Motor;
 use Illuminate\Http\Request;
 use App\Models\Kendaraan;
+use Illuminate\Support\Facades\DB;
 
 class KendaraanController extends Controller
 {
@@ -15,7 +16,7 @@ class KendaraanController extends Controller
         return response()->json($kendaraan);
     }
     
-    public function show($id)
+    public function detail($id)
     {
         $kendaraan = Kendaraan::with('motor', 'mobil')->find($id);
         return response()->json($kendaraan);
@@ -37,6 +38,11 @@ class KendaraanController extends Controller
             $motor->suspensi_motor = $request->input('suspensi_motor');
             $motor->transmisi_motor = $request->input('transmisi_motor');
             $motor->id_kendaraan = $kendaraan->_id;
+            $motor->nama = $kendaraan->nama;
+            $motor->tahun = $kendaraan->tahun;
+            $motor->warna = $kendaraan->warna;
+            $motor->harga = $kendaraan->harga;
+            $motor->stok = $kendaraan->stok;
             $motor->save();
             $kendaraan->jeniskendaraan()->save($motor);
         }
@@ -47,9 +53,14 @@ class KendaraanController extends Controller
             $mobil->kapasitas_mobil = $request->input('kapasitas_mobil');
             $mobil->tipe_mobil = $request->input('tipe_mobil');
             $mobil->id_kendaraan = $kendaraan->_id;
+            $mobil->nama = $kendaraan->nama;
+            $mobil->tahun = $kendaraan->tahun;
+            $mobil->warna = $kendaraan->warna;
+            $mobil->harga = $kendaraan->harga;
+            $mobil->stok = $kendaraan->stok;
             $mobil->save();
             $kendaraan->jeniskendaraan()->save($mobil);
-        }  
+        }
         
         return response()->json([
             'message' => 'Kendaraan berhasil ditambah.',
@@ -72,6 +83,11 @@ class KendaraanController extends Controller
             $motor->mesin_motor = $request->input('mesin_motor');
             $motor->suspensi_motor = $request->input('suspensi_motor');
             $motor->transmisi_motor = $request->input('transmisi_motor');
+            $motor->nama = $kendaraan->nama;
+            $motor->tahun = $kendaraan->tahun;
+            $motor->warna = $kendaraan->warna;
+            $motor->harga = $kendaraan->harga;
+            $motor->stok = $kendaraan->stok;
             $motor->save();
             $kendaraan->jeniskendaraan()->save($motor);
         }
@@ -81,17 +97,64 @@ class KendaraanController extends Controller
             $mobil->mesin_mobil = $request->input('mesin_mobil');
             $mobil->kapasitas_mobil = $request->input('kapasitas_mobil');
             $mobil->tipe_mobil = $request->input('tipe_mobil');
+            $mobil->nama = $kendaraan->nama;
+            $mobil->tahun = $kendaraan->tahun;
+            $mobil->warna = $kendaraan->warna;
+            $mobil->harga = $kendaraan->harga;
+            $mobil->stok = $kendaraan->stok;
             $mobil->save();
             $kendaraan->jeniskendaraan()->save($mobil);
         }
-
+        
         return response()->json($kendaraan);
     }
     
     public function destroy($id)
     {
         $kendaraan = Kendaraan::find($id);
+        
+        if (!$kendaraan) {
+            return response()->json(['message' => 'Kendaraan tidak ada.'], 404);
+        }
+
+        $mobil = Mobil::where('id_kendaraan', $id)->first();
+        if ($mobil) {
+            $mobil->delete();
+        }
+
+        $motor = Motor::where('id_kendaraan', $id)->first();
+        if ($motor) {
+            $motor->delete();
+        }
+
         $kendaraan->delete();
-        return response()->json(['message' => 'Kendaraan berhasil dihapus']);
+        return response()->json(['message' => 'Kendaraan dihapus.'], 200);
     }
+    
+    public function stock()
+    {
+        $motor = Motor::get();
+        $stokMotor = $motor->map(function ($kendaraan) {
+            return collect($kendaraan->toArray())
+            ->only(['stok'])
+            ->all();
+        });
+        
+        
+        $mobil = Mobil::get();
+        $stokMobil = $mobil->map(function ($kendaraan) {
+            return collect($kendaraan->toArray())
+            ->only(['stok'])
+            ->all();
+        });
+        
+        
+        return response()->json([
+            'stok motor'    => $stokMotor,
+            'stok mobil'    => $stokMobil,
+        ]);
+        
+        
+    }
+    
 }
